refactor: extract run-all ordering and request naming to lib/requestOps.ts
Move the dependency-chain topological ordering behind Run All and the
duplicate-checked alias assignment (nameRequest) out of App.svelte into
src/lib/requestOps.ts. runAllRequests takes the send function as a
parameter so the UI-store-updating manual send stays in App.svelte.
No behavior change.
